hash() now returns same value on 2-byte and 4-byte machines

+++ mach/proto/as/comm5.c
@@ -468,7 +468,7 @@ ok:
 hash(p)
 register char *p;
 {
-	register unsigned h;
+	register unsigned short h;
 	register c;
 
 	h = 0;
